feat(build-server): Move server build to Vite plugin
